Fix delete physic

+++ resources/web/wwi/X3dScene.js
@@ -136,7 +136,8 @@ export default class X3dScene {
 
   #deleteObject(id) {
     const object = WbWorld.instance.nodes.get('n' + id);
-    object.delete();
+    if (typeof object !== 'undefined')
+      object.delete();
 
     WbWorld.instance.robots.forEach((robot, i) => {
       if (robot.id === 'n' + id)
